tests: improve automated coverage
- airdrop: moved gov before airdrop (register keys);
- participationrewards: added ClaimType enum;
- airdrop: removed BeginBlocker;
- airdrop added EndBlocker to EndBlock;
- airdrop: improved error responses;
- airdrop: added keeper test suite;
- airdrop: added msg server tests;
- airdrop: added proposal handler tests;
- airdrop: updated proposal handler validation;
- airdrop: removed GetZoneDropAccount;
- participationrewards: ProtocolDatas confromed to enum (protobuf3 style);
- participationrewards: added more sentinel errors;
- participationrewards: refactored genesis validate (added protocoldata validation);
- participationrewards: added msg tests;
- participationrewards: added ValidateBasic (stubs);
- participationrewards: implemented preliminary validateProtocolData;

+++ x/participationrewards/keeper/keeper.go
@@ -95,7 +95,7 @@ func (k Keeper) GetAllocation(ctx sdk.Context, balance sdk.Coin, portion sdk.Dec
 
 func LoadSubmodules() map[int64]Submodule {
 	out := make(map[int64]Submodule, 0)
-	out[types.ClaimTypeOsmosisPool] = &OsmosisModule{}
-	out[types.ClaimTypeLiquidToken] = &LiquidTokensModule{}
+	out[int64(types.ClaimTypeOsmosisPool)] = &OsmosisModule{}
+	out[int64(types.ClaimTypeLiquidToken)] = &LiquidTokensModule{}
 	return out
 }
